Rename `accessors` folder
Accessors was an accurate name for this folder at the time, since it
contained only "load" operations, but we are going to want to create
similar helper functions for other types of operation (update, create,
and some day delete).
