+++ src/database/index.ts
@@ -1,5 +1,5 @@
-export * from './accessors';
 export * from './db';
 export * from './enrichers';
 export * from './migrate';
+export * from './operations';
 export * from './parameters';
