+++ src/database/operations/load/loadBaseFields.ts
@@ -1,5 +1,5 @@
-import { db } from '../db';
-import type { BaseField } from '../../types';
+import { db } from '../../db';
+import type { BaseField } from '../../../types';
 
 export const loadBaseFields = async (): Promise<BaseField[]> => (
   (await db.sql<BaseField>('baseFields.selectAll')).rows
