+++ src/database/operations/load/loadTableMetrics.ts
@@ -1,7 +1,7 @@
-import { db } from '../db';
-import { isTableMetrics } from '../../types';
-import { InternalValidationError } from '../../errors';
-import type { TableMetrics } from '../../types';
+import { db } from '../../db';
+import { isTableMetrics } from '../../../types';
+import { InternalValidationError } from '../../../errors';
+import type { TableMetrics } from '../../../types';
 
 export const loadTableMetrics = async (tableName: string): Promise<TableMetrics> => {
   const metricsQueryResult = await db.query<TableMetrics>(`
